code/requirements.txt
clams-python==1.0.3
